requirements.txt
aiofiles==23.2.1
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.7
click==8.1.8
click-default-group==1.2.4
comm==0.2.3
condense-json==0.1.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.8
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
llm==0.27.1
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.0.2
openai==2.16.0
orjson==3.11.5
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==15.0.0
ruff==0.15.11
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
soupsieve==2.8.3
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
stack-data==0.6.3
starlette==0.49.3
tabulate==0.9.0
tomlkit==0.12.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.39.0
wcwidth==0.6.0
websockets==12.0
zipp==3.23.0
